Move all `ble` examples to `ble` and convert to individual projects
