+++ examples/ble/embassy_ble/Cargo.toml
@@ -1,0 +1,87 @@
+[package]
+name    = "embassy-ble"
+version = "0.0.0"
+edition = "2024"
+publish = false
+
+[dependencies]
+bleps = { git = "https://github.com/bjoernQ/bleps", package = "bleps", rev = "a5148d8", features = [
+    "async",
+    "macros",
+] }
+cfg-if = "1.0.0"
+embassy-executor = { version = "0.7.0", features = ["task-arena-size-20480"] }
+esp-alloc = { path = "../../../esp-alloc" }
+esp-backtrace = { path = "../../../esp-backtrace", features = [
+    "panic-handler",
+    "println",
+] }
+esp-bootloader-esp-idf = { path = "../../../esp-bootloader-esp-idf" }
+esp-hal-embassy = { path = "../../../esp-hal-embassy" }
+esp-hal = { path = "../../../esp-hal", features = ["log-04", "unstable"] }
+esp-println = { path = "../../../esp-println", features = ["log-04"] }
+esp-radio-preempt-baremetal = { path = "../../../esp-radio-preempt-baremetal", features = [
+    "log-04",
+] }
+esp-radio = { path = "../../../esp-radio", features = [
+    "ble",
+    "log-04",
+    "unstable",
+] }
+static_cell = "2.1.0"
+
+[features]
+esp32 = [
+    "esp-backtrace/esp32",
+    "esp-bootloader-esp-idf/esp32",
+    "esp-hal-embassy/esp32",
+    "esp-hal/esp32",
+    "esp-radio-preempt-baremetal/esp32",
+    "esp-radio/esp32",
+]
+esp32c2 = [
+    "esp-backtrace/esp32c2",
+    "esp-bootloader-esp-idf/esp32c2",
+    "esp-hal-embassy/esp32c2",
+    "esp-hal/esp32c2",
+    "esp-radio-preempt-baremetal/esp32c2",
+    "esp-radio/esp32c2",
+]
+esp32c3 = [
+    "esp-backtrace/esp32c3",
+    "esp-bootloader-esp-idf/esp32c3",
+    "esp-hal-embassy/esp32c3",
+    "esp-hal/esp32c3",
+    "esp-radio-preempt-baremetal/esp32c3",
+    "esp-radio/esp32c3",
+]
+esp32c6 = [
+    "esp-backtrace/esp32c6",
+    "esp-bootloader-esp-idf/esp32c6",
+    "esp-hal-embassy/esp32c6",
+    "esp-hal/esp32c6",
+    "esp-radio-preempt-baremetal/esp32c6",
+    "esp-radio/esp32c6",
+]
+esp32h2 = [
+    "esp-backtrace/esp32h2",
+    "esp-bootloader-esp-idf/esp32h2",
+    "esp-hal-embassy/esp32h2",
+    "esp-hal/esp32h2",
+    "esp-radio-preempt-baremetal/esp32h2",
+    "esp-radio/esp32h2",
+]
+esp32s3 = [
+    "esp-backtrace/esp32s3",
+    "esp-bootloader-esp-idf/esp32s3",
+    "esp-hal-embassy/esp32s3",
+    "esp-hal/esp32s3",
+    "esp-radio-preempt-baremetal/esp32s3",
+    "esp-radio/esp32s3",
+]
+
+[profile.release]
+debug            = true
+debug-assertions = true
+lto              = "fat"
+codegen-units    = 1
